refactor(mcp): co-locate metadata tools with their modules
- Move object metadata tools to object-metadata/tools/
- Move field metadata tools to field-metadata/tools/
- Delete centralized metadata-tools module
- Update MCP layer to aggregate tools from individual modules
- Fix inconsistent return types in get operations
