+++ packages/twenty-server/src/engine/metadata-modules/field-metadata/field-metadata.module.ts
@@ -32,6 +32,7 @@ import { WorkspaceCacheStorageModule } from 'src/engine/workspace-cache-storage/
 import { WorkspaceMigrationRunnerModule } from 'src/engine/workspace-manager/workspace-migration-runner/workspace-migration-runner.module';
 import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-v2.module';
 import { FieldMetadataGraphqlApiExceptionInterceptor } from 'src/engine/metadata-modules/field-metadata/interceptors/field-metadata-graphql-api-exception.interceptor';
+import { FieldMetadataToolsFactory } from 'src/engine/metadata-modules/field-metadata/tools/field-metadata-tools.factory';
 
 import { FieldMetadataEntity } from './field-metadata.entity';
 
@@ -91,7 +92,11 @@ import { UpdateFieldInput } from './dtos/update-field.input';
       ],
     }),
   ],
-  providers: [FieldMetadataService, FieldMetadataResolver],
-  exports: [FieldMetadataService],
+  providers: [
+    FieldMetadataService,
+    FieldMetadataResolver,
+    FieldMetadataToolsFactory,
+  ],
+  exports: [FieldMetadataService, FieldMetadataToolsFactory],
 })
 export class FieldMetadataModule {}
